Added simple email formatting

+++ inventoryservice-persistence/src/main/resources/inventory.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `ITEMS`(
 
     `title`       VARCHAR PRIMARY KEY,
-    `price`       FLOAT NOT NULL,
+    `price`       DOUBLE NOT NULL,
     `image`      BLOB NOT NULL,
     `quantity`    INTEGER NOT NULL,
     `description` VARCHAR NOT NULL
